setting extension node selectors moved
